% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initials.R
\name{abbreviate}
\alias{abbreviate}
\title{Abbreviate middle names}
\usage{
abbreviate(string, collapse)
}
\arguments{
\item{string}{Character. A character vector with the names}

\item{collapse}{Character. A string that will be used to separate names}
}
\value{
Returns a character vector with one element.
}
\description{
Abbreviates multiple words to first letters
}
\examples{
tenzing:::abbreviate("Franz Jude Wayne", collapse = "")
}
