% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initials.R
\name{abbreviate_middle_names_df}
\alias{abbreviate_middle_names_df}
\title{Abbreviate middle names in a dataframe}
\usage{
abbreviate_middle_names_df(contributors_table)
}
\arguments{
\item{contributors_table}{the imported contributors_table}
}
\value{
The function returns a dataframe with abbreviated middle
names.
}
\description{
The function calls the \code{\link{abbreviate}} function to
abbreviate middle names in the `Middle name` variable in a
dataframe if they are present. The function requires a valid
`contributors_table` as an input to work.
}
