% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_config.R
\name{add_analytics_script}
\alias{add_analytics_script}
\title{Add analytics script}
\usage{
add_analytics_script()
}
\value{
HTML tags for analytics script
}
\description{
Adds Matomo analytics tracking script.
}
