% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_config.R
\name{add_custom_js_files}
\alias{add_custom_js_files}
\title{Add custom JavaScript files}
\usage{
add_custom_js_files()
}
\value{
HTML tags for custom JavaScript files
}
\description{
Adds application-specific JavaScript files.
}
\keyword{internal}
