% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_config.R
\name{add_js_libraries}
\alias{add_js_libraries}
\title{Add external JavaScript libraries}
\usage{
add_js_libraries()
}
\value{
HTML tags for JavaScript libraries
}
\description{
Adds SweetAlert2 and Prism.js libraries.
}
