% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_affiliation_columns}
\alias{check_affiliation_columns}
\title{Check affiliation columns}
\usage{
check_affiliation_columns(contributors_table)
}
\arguments{
\item{contributors_table}{dataframe. Contributors table read with the \code{tenzing::read_contributors_table} function}
}
\description{
Affiliation columns can be named in multiple ways. In the first version of tenzing only two affiliation were allowed per contributor
and the required names for the columns were \verb{Primary affiliation} and \verb{Secondary affiliation}. In the new \code{deluxe} version of the spreadsheet
any number of affiliation columns can be created as long as they follow the \verb{Affiliation \{number\}} naming convention.
}
