% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_affiliation_consistency}
\alias{check_affiliation_consistency}
\title{Check affiliation columns for consistency}
\usage{
check_affiliation_consistency(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe containing the contributors' table.}
}
\value{
A list containing \code{valid} (TRUE/FALSE) and a \code{message} (if applicable).
}
\description{
This function checks if both legacy (\verb{Primary affiliation} and \verb{Secondary affiliation})
and \verb{Affiliation \{n\}} columns are present in the contributors_table. If both are present,
it raises a warning to suggest using only one format to ensure consistent results.
}
