% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_author_acknowledgee_values}
\alias{check_author_acknowledgee_values}
\title{Check allowed values in Author/Acknowledgee}
\usage{
check_author_acknowledgee_values(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe of contributors.}
}
\value{
list(type = "success"|"warning", message = \if{html}{\out{<chr>}})
}
\description{
Verifies that all entries in \code{Author/Acknowledgee} belong to the allowed set:
"Author", "Acknowledgment only", "Don't agree to be named".
}
