% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_coi_column_rename}
\alias{check_coi_column_rename}
\title{Check for Old Conflict of Interest Column Name}
\usage{
check_coi_column_rename(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe containing the contributors' information.}
}
\value{
A list containing:
\item{type}{Type of validation result: "success" or "warning".}
\item{message}{An informative message indicating that the column name needs to be updated.}
}
\description{
This function checks if the old column name "Conflict of interest" is present
in the contributors table. If found, it returns a warning with instructions
to update the column name to "Declares".
}
