% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_cols}
\alias{check_cols}
\title{Check column names in the contributors_table}
\usage{
check_cols(contributors_table)
}
\arguments{
\item{contributors_table}{dataframe, filled out contributors_table}
}
\description{
This helper function for the validation checks whether all the necessary columns in the contributors_table are available.
}
