% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_corresponding_non_author}
\alias{check_corresponding_non_author}
\title{Check that only Authors are marked as Corresponding}
\usage{
check_corresponding_non_author(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe of contributors.}
}
\value{
list(type = "success"|"warning", message = \if{html}{\out{<chr>}})
}
\description{
Warns if any row is marked \verb{Corresponding author? == TRUE} while
\code{Author/Acknowledgee} is not "Author".
}
