% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_infosheet.R
\name{check_duplicate_initials}
\alias{check_duplicate_initials}
\title{Check for same initials}
\usage{
check_duplicate_initials(infosheet)
}
\arguments{
\item{infosheet}{the imported infosheet}
}
\value{
The function returns
}
\description{
This function checks the infosheet for duplicate initials, and
issues a warning that the surnames will be used to differentiate
between the users.
}
