% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_missing_author_acknowledgee}
\alias{check_missing_author_acknowledgee}
\title{Check missing Author/Acknowledgee where names are present}
\usage{
check_missing_author_acknowledgee(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe of contributors.}
}
\value{
list(type = "success"|"warning", message = \if{html}{\out{<chr>}})
}
\description{
Warns if \code{Author/Acknowledgee} is missing for rows that have a name
(Firstname or Surname present).
}
