% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_missing_corresponding}
\alias{check_missing_corresponding}
\title{Check if the corresponding author is missing}
\usage{
check_missing_corresponding(contributors_table)
}
\description{
Check if the corresponding author is missing
}
