% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_missing_email}
\alias{check_missing_email}
\title{Check if email is provided for the corresponding author}
\usage{
check_missing_email(contributors_table)
}
\description{
Check if email is provided for the corresponding author
}
