% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_missing_orcid}
\alias{check_missing_orcid}
\title{Check for Missing ORCID IDs}
\usage{
check_missing_orcid(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe containing the contributors' information.}
}
\value{
A list containing:
\item{type}{Type of validation result: "success" or "warning".}
\item{message}{An informative message indicating the row numbers with missing ORCID IDs, if any.}
}
\description{
This function checks for missing values in the \verb{ORCID iD} column of the
\code{contributors_table} and returns a warning if any ORCID IDs are missing.
}
\details{
If the \verb{ORCID iD} column does not exist, the function returns a success message
since ORCID IDs are optional.
}
