% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_outputs}
\alias{check_outputs}
\title{Check column names in the contributors_table}
\usage{
check_outputs(contributors_table)
}
\arguments{
\item{contributors_table}{dataframe, filled out contributors_table}
}
\value{
The function returns a list that contains a nested list for each output that contains: boolean for whether the output is enabled,
and vector of missing column names.
}
\description{
This helper function for the validation checks whether all the necessary columns in the contributors_table are available for each
output separately.
}
