% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_roles_present}
\alias{check_roles_present}
\title{Check for Presence of Assigned CRediT Roles}
\usage{
check_roles_present(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe containing the contributors' information.}

\item{context}{Optional named list providing contextual information.}
}
\value{
A list containing:
\item{type}{Type of validation result: "success" or "error".}
\item{message}{A descriptive validation message.}
}
\description{
This function verifies that at least one CRediT taxonomy role
is checked for any contributor in the filtered dataset.
}
\details{
This validation is equivalent to the internal check used in
\code{print_credit_roles()} to ensure that contributors have at least one
CRediT category marked as TRUE.
}
\examples{
check_roles_present(filtered_tbl)

}
