% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_contributors_table.R
\name{clean_contributors_table}
\alias{clean_contributors_table}
\title{Delete empty rows of the contributors_table}
\usage{
clean_contributors_table(contributors_table)
}
\arguments{
\item{contributors_table}{the imported contributors_table}
}
\value{
The function returns the contributors_table
without empty additional rows.
}
\description{
The function deletes any additional rows where all
of the name columns are empty.
}
