% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_infosheet.R
\name{clean_infosheet}
\alias{clean_infosheet}
\title{Delete empty rows of the infosheet}
\usage{
clean_infosheet(infosheet)
}
\arguments{
\item{infosheet}{the imported infosheet}
}
\value{
The function returns the infosheet without empty additional rows.
}
\description{
The function deletes any additional rows where all
of the name columns are empty.
}
