% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_contrib_affil.R
\name{contrib_affil_print}
\alias{contrib_affil_print}
\title{Generate contributors' affiliations}
\usage{
contrib_affil_print(infosheet)
}
\arguments{
\item{infosheet}{validated infosheet}
}
\value{
The output is a list containing the contributors' name in the
  the order defined by the \code{Order in publication} column of the
  infosheet in the \code{contrib} character vector, and the adherent
  affiliations in the same order in the \code{affil} character vector.
}
\description{
The function generates rmarkdown formatted contributors' affiliation text from
an infosheet validated with the \code{\link{validate_infosheet}} function. The 
infosheet must be based on the \code{\link{infosheet_template}}.
}
\section{Warning}{

The function is primarily developed to be the part of a shiny app. As the
  validation is handled inside of the app separately, the function can
  break with non-informative errors if running locally without first
  validating it.
}

\examples{
validate_infosheet(infosheet = infosheet_template)
print_contrib_affil(infosheet = infosheet_template)
}
\seealso{
Other \code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}: 
\code{\link{print_yaml}()},
\code{\link{roles_readable_print}()},
\code{\link{xml_print}()}
}
\concept{\code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}}
