% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{contrib_group}
\alias{contrib_group}
\title{Function to create the XML document}
\usage{
contrib_group(contrib_data)
}
\arguments{
\item{contrib_data}{preformatted contributors_table}
}
\value{
The function returns an XML document from the contributors
table formatted according the JATS 1.2 DTD specifications.
}
\description{
This function generates an XML document from the preformatted
\code{contributors_table}.
}
