% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{generate_affiliations}
\alias{generate_affiliations}
\title{Generate affiliation XML nodes}
\usage{
generate_affiliations(affiliation_data)
}
\arguments{
\item{affiliation_data}{vector of unique affiliations}
}
\description{
Generate affiliation XML nodes
}
\keyword{internal}
