% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_configs.R
\name{generate_all_configs}
\alias{generate_all_configs}
\title{Generate all standard validation configurations}
\usage{
generate_all_configs(config_dir = "inst/config")
}
\arguments{
\item{config_dir}{Directory where config files should be saved}
}
\description{
Generate all standard validation configurations
}
