% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{generate_author_notes}
\alias{generate_author_notes}
\title{Generate author-notes XML node}
\usage{
generate_author_notes(contributors_table)
}
\arguments{
\item{contributors_table}{contributors table}
}
\description{
Generate author-notes XML node
}
\keyword{internal}
