% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{generate_contrib_group}
\alias{generate_contrib_group}
\title{Generate contrib-group XML node}
\usage{
generate_contrib_group(contrib_data, affiliation_data, authors_only)
}
\arguments{
\item{contrib_data}{preprocessed contributors table with affiliations}

\item{affiliation_data}{vector of unique affiliations}

\item{authors_only}{full authors table for metadata}
}
\description{
Generate contrib-group XML node
}
\keyword{internal}
