% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{generate_funding_group}
\alias{generate_funding_group}
\title{Generate funding-group XML node}
\usage{
generate_funding_group(contributors_table)
}
\arguments{
\item{contributors_table}{contributors table}
}
\description{
Generate funding-group XML node
}
\keyword{internal}
