% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenzing_utils_general.R
\name{glue_oxford_collapse}
\alias{glue_oxford_collapse}
\title{Collapse a character vector with oxford comma}
\usage{
glue_oxford_collapse(x)
}
\arguments{
\item{x}{character, the vector to be collapsed}
}
\value{
The function returns a vector of length 1.
}
\description{
Collapses a character vector into a length 1 vector,
by using ", " as a separator and adding the oxford comma
if there original character vector length is longer than 3.
The function is from the cli package: https://github.com/jonocarroll/cli/blob/2d3fbc4b41327df82df1102cdfc0a5c99822809b/R/inline.R
}
\keyword{internal}
