% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infosheet_template.R
\docType{data}
\name{infosheet_template}
\alias{infosheet_template}
\title{Template infosheet}
\format{A dataframe with 3 rows and 22 variables:
\describe{
  \item{'Order in publication'}{numeric, used to order the contributors for all outputs}
  \item{Firstname}{character, first name of the author, must be provided}
  \item{'Middle name'}{character, middle name of the author, blank if not applicable}
  \item{Surname}{character, surname of the author, must be provided}
  \item{Conceptualization}{logical, CREdiT role \url{}}
  \item{'Data Curation'}{logical, CREdiT role \url{}}
  \item{'Formal Analysis'}{logical, CREdiT role \url{}}
  \item{'Funding Acquisition'}{logical, CREdiT role \url{}}
  \item{Investigation}{logical, CREdiT role \url{}}
  \item{Methodology}{logical, CREdiT role \url{}}
  \item{'Project Administration'}{logical, CREdiT role \url{}}
  \item{Resources}{logical, CREdiT role \url{}}
  \item{Software}{logical, CREdiT role \url{}}
  \item{Supervision}{logical, CREdiT role \url{}}
  \item{Validation}{logical, CREdiT role \url{}}
  \item{Visualization}{logical, CREdiT role \url{}}
  \item{'Writing - Original Draft Preparation'}{logical, CREdiT role \url{}}
  \item{'Writing - Review & Editing'}{logical, CREdiT role \url{}}
  \item{'Email address'}{character, email address of the correspondign author, optional}
  \item{'Primary affiliation'}{character, primary affiliation of the contributor}
  \item{'Secondary affiliation'}{character, secondary affiliation of the contributor, blank if not applicable}
  \item{'Corresponding author?'}{logical, TRUE for contributor who is the corresponding author}
}}
\source{
\url{https://docs.google.com/spreadsheets/d/1Gl0cwqN_nTsdFH9yhSvi9NypBfDCEhViGq4A3MnBrG8/edit?usp=sharing}
}
\usage{
infosheet_template
}
\description{
Dataset that represents the data structure that the tenzing
package functions require to run without error. The dataset
is filled with example values and based on the online spreadsheet
provided in the shiny app.
}
\section{Remark}{

Each row contains the contributorship information for
one contributor. In the template there are 3 contributors
added as an example but there is no limit to the number
of contributors while using the package.
}

\keyword{datasets}
