% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{js_add_tooltip}
\alias{js_add_tooltip}
\title{Add tooltip to element}
\usage{
js_add_tooltip(selector, message)
}
\arguments{
\item{selector}{CSS selector for element}

\item{message}{Tooltip message}
}
\description{
Add tooltip to element
}
