% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{js_disable_buttons}
\alias{js_disable_buttons}
\title{Disable buttons by selector}
\usage{
js_disable_buttons(selector)
}
\arguments{
\item{selector}{CSS selector for buttons to disable}
}
\description{
Disable buttons by selector
}
\keyword{internal}
