% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{js_error_alert}
\alias{js_error_alert}
\title{Show error alert}
\usage{
js_error_alert(error, warning = "")
}
\arguments{
\item{error}{Error message}

\item{warning}{Warning message (optional, defaults to empty string)}
}
\description{
Show error alert
}
\keyword{internal}
