% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{js_remove_tooltip}
\alias{js_remove_tooltip}
\title{Remove tooltip from element}
\usage{
js_remove_tooltip(selector)
}
\arguments{
\item{selector}{CSS selector for element}
}
\description{
Remove tooltip from element
}
\keyword{internal}
