% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{js_success_alert}
\alias{js_success_alert}
\title{Show success alert}
\usage{
js_success_alert(message = "")
}
\arguments{
\item{message}{Success message (optional, defaults to empty string)}
}
\description{
Show success alert
}
