% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{js_update_card_styles}
\alias{js_update_card_styles}
\title{Update validation card styles}
\usage{
js_update_card_styles(card_id, header_text_id, text_color, border_color)
}
\arguments{
\item{card_id}{Card element ID}

\item{header_text_id}{Header text element ID}

\item{text_color}{Text color hex code}

\item{border_color}{Border color hex code}
}
\description{
Update validation card styles
}
