% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about_modal.R
\name{mod_about_modal_ui}
\alias{mod_about_modal_ui}
\alias{mod_about_modal_server}
\title{mod_about_modal_ui and mod_about_modal_server}
\usage{
mod_about_modal_ui(id)

mod_about_modal_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
