% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contribs_affiliation_page.R
\name{mod_contribs_affiliation_page_ui}
\alias{mod_contribs_affiliation_page_ui}
\alias{mod_contribs_affiliation_page_server}
\title{mod_contribs_affiliation_page_ui and mod_contribs_affiliation_page_server}
\usage{
mod_contribs_affiliation_page_ui(id)

mod_contribs_affiliation_page_server(id, input_data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
