% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_create_table.R
\name{mod_create_table_ui}
\alias{mod_create_table_ui}
\alias{mod_create_table_server}
\title{mod_create_table_ui and mod_create_table_server}
\usage{
mod_create_table_ui(id)

mod_create_table_server(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
