% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_human_readable_report.R
\name{mod_human_readable_report_ui}
\alias{mod_human_readable_report_ui}
\alias{mod_human_readable_report_server}
\title{mod_human_readable_report_ui and mod_human_readable_report_server}
\usage{
mod_human_readable_report_ui(id)

mod_human_readable_report_server(id, input_data, valid_infosheet)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
