% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_read_spreadsheet.R
\name{mod_read_spreadsheet_ui}
\alias{mod_read_spreadsheet_ui}
\alias{mod_read_spreadsheet_server}
\title{mod_read_spreadsheet_ui and mod_read_spreadsheet_server}
\usage{
mod_read_spreadsheet_ui(id)

mod_read_spreadsheet_server(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
