% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_show_spreadsheet.R
\name{mod_show_spreadsheet_ui}
\alias{mod_show_spreadsheet_ui}
\alias{mod_show_spreadsheet_server}
\title{mod_show_spreadsheet_ui and mod_show_spreadsheet_server}
\usage{
mod_show_spreadsheet_ui(id)

mod_show_spreadsheet_server(id, input_data, uploaded)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
