% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_show_yaml.R
\name{mod_show_yaml_ui}
\alias{mod_show_yaml_ui}
\alias{mod_show_yaml_server}
\title{mod_show_yaml_ui and mod_show_yaml_server}
\usage{
mod_show_yaml_ui(id)

mod_show_yaml_server(id, input_data, valid_infosheet)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
