% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_xml_report.R
\name{mod_xml_report_ui}
\alias{mod_xml_report_ui}
\alias{mod_xml_report_server}
\title{mod_xml_report_ui and mod_xml_report_server}
\usage{
mod_xml_report_ui(id)

mod_xml_report_server(id, input_data, submit)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
