% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleServer.R
\name{moduleServer}
\alias{moduleServer}
\title{Wrapper function around \code{shiny::callModule}}
\usage{
moduleServer(id, module)
}
\arguments{
\item{id}{Unique character id of the module}

\item{module}{Name of the module server function}
}
\description{
The function dodge the explicit usage of \code{callModule} on
the server side. This style will be the default for \code{shiny}
from version 1.5.0. To be deleted after update.
}
\section{Source}{

\url{https://mastering-shiny.org/scaling-modules.html}
}

