% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_conflict_statement.R
\name{print_conflict_statement}
\alias{print_conflict_statement}
\title{Generate human readable report of the conflict of interest statements}
\usage{
print_conflict_statement(contributors_table, initials = FALSE)
}
\arguments{
\item{contributors_table}{validated contributors_table}

\item{initials}{Logical. If true initials will be included instead of full
names in the output}
}
\value{
The function returns a string.
}
\description{
The functions generates the conflict of interest statement section of the manuscript.
The output is generated from an contributors_table validated with
the \code{\link[=validate_contributors_table]{validate_contributors_table()}} function.
The contributors_table must be based on the \code{\link[=contributors_table_template]{contributors_table_template()}}.
}
\examples{
example_contributors_table <- read_contributors_table(
contributors_table = system.file("extdata",
"contributors_table_example.csv", package = "tenzing", mustWork = TRUE))
validate_contributors_table(contributors_table = example_contributors_table)
print_conflict_statement(contributors_table = example_contributors_table, initials = FALSE)

}
\seealso{
Other output functions: 
\code{\link{print_credit_roles}()},
\code{\link{print_funding}()},
\code{\link{print_title_page}()},
\code{\link{print_xml}()},
\code{\link{print_yaml}()}
}
\concept{output functions}
