% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_credit_roles.R
\name{print_credit_roles}
\alias{print_credit_roles}
\title{Generate report of the contributions with CRedit}
\usage{
print_credit_roles(
  contributors_table,
  text_format = "rmd",
  initials = FALSE,
  order_by = "role"
)
}
\arguments{
\item{contributors_table}{Tibble. Validated contributors_table}

\item{text_format}{Character. Formatting of the returned string. Possible values: "rmd", "html", "raw".
"rmd" by default.}

\item{initials}{Logical. If true initials will be included instead of full
names in the output}

\item{order_by}{Character. Whether the contributing authors listed for each role ("role"), or
the roles are listed after the name of each contributor ("contributor").}
}
\value{
The function returns a string containing the CRediT roles
  with the contributors listed for each role they partake in.
}
\description{
The function generates rmarkdown formatted text of the contributions according
to the CRediT taxonomy. The output is generated from an `contributors_table` validated with
the \code{\link{validate_contributors_table}} function. The `contributors_table` must be based on the
\code{\link{contributors_table_template}}. The function can return the output string as
rmarkdown or html formatted text or without any formatting.
}
\section{Warning}{

The function is primarily developed to be the part of a shiny app. As the
  validation is handled inside of the app separately, the function can
  break with non-informative errors if running locally without first
  validating it.
}

\examples{
example_contributors_table <- read_contributors_table(
contributors_table = system.file("extdata",
"contributors_table_example.csv", package = "tenzing", mustWork = TRUE))
validate_contributors_table(contributors_table = example_contributors_table)
print_credit_roles(contributors_table = example_contributors_table)

}
\seealso{
Other output functions: 
\code{\link{print_funding}()},
\code{\link{print_title_page}()},
\code{\link{print_xml}()},
\code{\link{print_yaml}()}
}
\concept{output functions}
