% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_funding.R
\name{print_funding}
\alias{print_funding}
\title{Generate human readable report of the funding information}
\usage{
print_funding(contributors_table, initials = FALSE)
}
\arguments{
\item{contributors_table}{validated contributors_table}

\item{initials}{Logical. If true initials will be included instead of full
names in the output}
}
\value{
The function returns a string.
}
\description{
The functions generates the funding information section of the manuscript.
The output is generated from an contributors_table based on the \code{\link[=contributors_table_template]{contributors_table_template()}}.
}
\examples{
example_contributors_table <- read_contributors_table(
contributors_table = system.file("extdata",
"contributors_table_example.csv", package = "tenzing", mustWork = TRUE))
print_funding(contributors_table = example_contributors_table, initials = FALSE)

}
\seealso{
Other output functions: 
\code{\link{print_conflict_statement}()},
\code{\link{print_credit_roles}()},
\code{\link{print_title_page}()},
\code{\link{print_xml}()},
\code{\link{print_yaml}()}
}
\concept{output functions}
