% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_grant.R
\name{print_grant}
\alias{print_grant}
\title{Generate human readable report of the grant information}
\usage{
print_grant(infosheet, initials = FALSE)
}
\arguments{
\item{infosheet}{validated infosheet}

\item{initials}{Logical. If true initials will be included instead of full
names in the output}
}
\value{
The function returns a string.
}
\description{
The functions generates a human readable text about the funding
information of the contributors. The output is generated from an
infosheet validated with the \code{\link{validate_infosheet}} function.
The infosheet must be based on the \code{\link{infosheet_template}}.
}
\seealso{
Other output functions: 
\code{\link{print_contrib_affil}()},
\code{\link{print_roles_readable}()},
\code{\link{print_xml}()},
\code{\link{print_yaml}()}
}
\concept{output functions}
