% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_roles_readable.R
\name{print_roles_readable}
\alias{print_roles_readable}
\title{Generate human readable report of the contributions}
\usage{
print_roles_readable(infosheet, text_format = "rmd", initials = FALSE)
}
\arguments{
\item{infosheet}{validated infosheet}

\item{text_format}{formatting of the returned string. Possible values: "rmd", "html", "raw".
"rmd" by default.}

\item{initials}{Logical. If true initials will be included instead of full
names in the output}
}
\value{
The function returns a string containing the CRediT roles
  with the contributors listed for each role they partake in.
}
\description{
The function generates rmarkdown formatted text of the contributions according
to the CRediT taxonomy. The output is generated from an infosheet validated with
the \code{\link{validate_infosheet}} function. The infosheet must be based on the
\code{\link{infosheet_template}}. The function can return the output string as
rmarkdown or html formatted text or without any formatting.
}
\section{Warning}{

The function is primarily developed to be the part of a shiny app. As the
  validation is handled inside of the app separately, the function can
  break with non-informative errors if running locally without first
  validating it.
}

\examples{
validate_infosheet(infosheet = infosheet_template)
print_roles_readable(infosheet = infosheet_template)
}
\seealso{
Other output functions: 
\code{\link{print_contrib_affil}()},
\code{\link{print_grant}()},
\code{\link{print_xml}()},
\code{\link{print_yaml}()}
}
\concept{output functions}
