% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_contributors_table.R
\name{read_contributors_table}
\alias{read_contributors_table}
\title{Read the filled out contributors_table}
\usage{
read_contributors_table(contributors_table_path)
}
\arguments{
\item{contributors_table_path}{the full path to the file with extension}
}
\value{
The function returns the contributors table as
a dataframe.
}
\description{
This function reads the \code{contributors_table} given the path if the
file is a csv, tsv or an xlsx. The function can read
googlesheets if share url is provided or local files if
path to the local folder is provided.
}
\section{Warning}{

If the file is an xlsx the function only reads the first sheet.
}

