% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_infosheet.R
\name{read_infosheet}
\alias{read_infosheet}
\title{Read the filled out infosheet}
\usage{
read_infosheet(infosheet_path)
}
\arguments{
\item{infosheet_path}{the full path to the file with extension}
}
\description{
Then function reads the infosheet given the path if the
file is a csv, tsv or an xlsx.
}
\section{Warning}{

If the file is an xlsx the function only reads the first sheet.
}

