% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Wrapper around \code{rmarkdown::render}}
\usage{
render_report(input, output, format, params)
}
\arguments{
\item{input}{path of the input RMD skeleton}

\item{output}{path of the rendered output file}

\item{format}{the extension of the output file}

\item{params}{list of parameters that will be passed to the RMD}
}
\description{
A wrapper function around \code{rmarkdown::render} in order
to call it in a different process in the modules.
}
\section{Source}{

The function is based on the suggestion of Hadley Wickham \url{https://mastering-shiny.org/action-transfer.html}.
}

\seealso{
\code{\link[rmarkdown]{render}}
}
