% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_roles_readable.R
\name{roles_readable_print}
\alias{roles_readable_print}
\title{Generate human readable report of the contributions}
\usage{
roles_readable_print(infosheet)
}
\arguments{
\item{infosheet}{validated infosheet

#' @return The function returns a character vector of the CRediT roles
  with the contributors listed for each role they partake in. The returned
  text is rmarkdown formatted.}
}
\description{
The function generates rmarkdown formatted text of the contributions according
to the CRediT taxonomy. The output is generated from an infosheet validated with
the \code{\link{validate_infosheet}} function. The infosheet must be based on the
\code{\link{infosheet_template}}.
}
\details{
#' @section Warning:
The function is primarily developed to be the part of a shiny app. As the
  validation is handled inside of the app separately, the function can
  break with non-informative errors if running locally without first
  validating it.
}
\examples{
validate_infosheet(infosheet = infosheet_template)
print_roles_readable(infosheet = infosheet_template)
}
\seealso{
Other \code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}: 
\code{\link{contrib_affil_print}()},
\code{\link{print_yaml}()},
\code{\link{xml_print}()}
}
\concept{\code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}}
