% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Allows user to pass global arguments to the app on runtime}
}
\description{
This function allows users to run the application
locally from their computer.
}
