% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_config.R
\name{setup_resource_paths}
\alias{setup_resource_paths}
\title{Setup resource paths}
\usage{
setup_resource_paths()
}
\value{
No return value, called for side effects
}
\description{
Sets up resource paths for www directory.
}
\keyword{internal}
