% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fragments.R
\name{ui_citation_section}
\alias{ui_citation_section}
\title{Create citation section}
\usage{
ui_citation_section()
}
\value{
HTML tag for citation section
}
\description{
Citation HTML is kept inline for easy editing.
}
