% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fragments.R
\name{ui_help_icon}
\alias{ui_help_icon}
\title{Create a help icon widget}
\usage{
ui_help_icon(help_text)
}
\arguments{
\item{help_text}{Tooltip text to display}
}
\value{
HTML tag for help icon
}
\description{
Create a help icon widget
}
