% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fragments.R
\name{ui_privacy_notice}
\alias{ui_privacy_notice}
\title{Create privacy notice section}
\usage{
ui_privacy_notice()
}
\value{
HTML tag for privacy notice
}
\description{
Privacy notice HTML is kept inline for easy editing.
}
\keyword{internal}
