% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenzing_utils_general.R
\name{unnamed_message}
\alias{unnamed_message}
\title{Return messages as unnamed vector for js handler}
\usage{
unnamed_message(x, y)
}
\arguments{
\item{x}{The result of the validation}

\item{y}{The type of validation message}
}
\value{
Returns a character vector with one element.
}
\description{
The function filters the results of the validation of the contributors_table template
by type. Then the message elements are pulled and collapsed into one element.
}
