% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_contributors_table.R
\name{validate_contributors_table}
\alias{validate_contributors_table}
\title{Validating the contributors table}
\usage{
validate_contributors_table(contributors_table, config_path)
}
\arguments{
\item{contributors_table}{dataframe, filled out contributors_table}

\item{config_path}{character, file path to validation configuration file}
}
\value{
The function returns a list for each checked statement. Each list contains
a \code{type} vector that stores whether the statement passed the check "success"
or failed "warning" or "error", and a \code{message} vector that contains information
about the nature of the check.
}
\description{
This function validates the \code{contributors_table} provided to it by checking whether the
provided \code{contributors_table} is compatible with the \code{\link[=contributors_table_template]{contributors_table_template()}}. The function
early escapes only if the provided \code{contributors_table} is not a dataframe, the variable
names that are present in the \code{contributors_table_template} is missing, or if the
\code{contributors_table} is empty.
}
\section{The function checks the following statements}{

\itemize{
\item error, the provided contributors_table is not a dataframe
\item error, none of the outputs can be created based the provided contributors_table due to missing columns
\item error, the provided contributors_table is empty
}
}

