% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\name{Validator}
\alias{Validator}
\title{Validator Class for Contributors Table}
\description{
Validator Class for Contributors Table

Validator Class for Contributors Table
}
\details{
The \code{Validator} class runs a set of user-defined validation functions on
a contributors table. It allows for configuring which validations should
be executed, handling dependencies between validations, and storing results.

This class is used in conjunction with the \code{ValidateOutput} class to
apply both column and general table validations.
}
\section{Configurable Validations}{

The class dynamically loads validation functions, allowing users to add
custom checks. By default, the predefined functions in \code{validate_helpers.R}
are available.
}

\section{Dependencies}{

Some validations depend on the presence of specific columns or successful
execution of other validations. These dependencies are defined in a YAML
config file.
}

\section{YAML Configuration}{

The validator reads a YAML configuration file (e.g., \code{inst/config/validator_example.yaml}),
which specifies:
\itemize{
\item The \strong{validations to run}.
\item Any \strong{dependencies} between them.
}

Example:
\preformatted{
validation_config:
  validations:
    - name: check_missing_corresponding
      dependencies:
        - '"Corresponding author?" %in% colnames(contributors_table)'
    - name: check_missing_email
      dependencies:
        - '"Corresponding author?" %in% colnames(contributors_table)'
        - 'self$results[["check_missing_corresponding"]]$type == "success"'
        - '"Email address" %in% colnames(contributors_table)'
}
}

\section{Usage}{

\preformatted{
# Create a Validator instance
validator <- Validator$new()

# Configure which validations should run
validator$setup_validator(validation_config)

# Run the validations on a contributors table
results <- validator$run_validations(contributors_table)
}
}

\seealso{
\code{\link{ValidateOutput}} which integrates this class for validation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{validations}}{A list of validation functions dynamically loaded from \code{validate_helpers.R} or custom functions.}

\item{\code{dependencies}}{A list of validation dependencies.}

\item{\code{results}}{Stores the results of executed validations.}

\item{\code{specified_validations}}{The subset of validations to execute, defined in the YAML config.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Validator-new}{\code{Validator$new()}}
\item \href{#method-Validator-add_dependency}{\code{Validator$add_dependency()}}
\item \href{#method-Validator-setup_validator}{\code{Validator$setup_validator()}}
\item \href{#method-Validator-run_validations}{\code{Validator$run_validations()}}
\item \href{#method-Validator-should_run}{\code{Validator$should_run()}}
\item \href{#method-Validator-clone}{\code{Validator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Validator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Validator-new}{}}}
\subsection{Method \code{new()}}{
Initializes the \code{Validator} class.
Loads predefined validations from \code{validate_helpers.R} and allows for adding custom validations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Validator-add_dependency"></a>}}
\if{latex}{\out{\hypertarget{method-Validator-add_dependency}{}}}
\subsection{Method \code{add_dependency()}}{
Adds dependencies for a validation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$add_dependency(validation_name, conditions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{validation_name}}{The name of the validation.}

\item{\code{conditions}}{A list of conditions that must be met before running the validation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Validator-setup_validator"></a>}}
\if{latex}{\out{\hypertarget{method-Validator-setup_validator}{}}}
\subsection{Method \code{setup_validator()}}{
Configures the validator with the subset of validations to execute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$setup_validator(validation_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{validation_config}}{A list defining which validations to run and their dependencies.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Validator-run_validations"></a>}}
\if{latex}{\out{\hypertarget{method-Validator-run_validations}{}}}
\subsection{Method \code{run_validations()}}{
Runs the specified validations on the provided contributors table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$run_validations(contributors_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contributors_table}}{A dataframe containing contributor data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of validation results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Validator-should_run"></a>}}
\if{latex}{\out{\hypertarget{method-Validator-should_run}{}}}
\subsection{Method \code{should_run()}}{
Determines whether a validation should be executed based on dependencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$should_run(validation_name, contributors_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{validation_name}}{The validation function name.}

\item{\code{contributors_table}}{The dataframe containing contributor data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the validation should run, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Validator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Validator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
