% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{xml_print}
\alias{xml_print}
\title{Generate an XML document of the contributions}
\usage{
xml_print(infosheet)
}
\arguments{
\item{infosheet}{validated infosheet

#' @return The function returns an xml nodeset containing the contributors
  listed for each CRediT role they partake in.}
}
\description{
The function generates an XML nodeset that contains the contributors' name,
affiliation, and their CRediT roles with a structure outlined in the
JATS 1.2 DTD specifications (eLife). The output is generated from an 
infosheet validated with the \code{\link{validate_infosheet}} function.
The infosheet must be based on the \code{\link{infosheet_template}}.
}
\details{
#' @section Warning:
The function is primarily developed to be the part of a shiny app. As the
  validation is handled inside of the app separately, the function can
  break with non-informative errors if running locally without first
  validating it.
}
\examples{
validate_infosheet(infosheet = infosheet_template)
print_xml(infosheet = infosheet_template)
}
\seealso{
Other \code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}: 
\code{\link{contrib_affil_print}()},
\code{\link{print_yaml}()},
\code{\link{roles_readable_print}()}
}
\concept{\code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}}
