% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenzing_utils_abbreviate.R
\name{abbreviate}
\alias{abbreviate}
\title{Abbreviate middle names}
\usage{
abbreviate(x)
}
\arguments{
\item{x}{Character. A character vector with one element.}
}
\value{
Returns a character vector with one element.
}
\description{
Abbreviates multiple words to first letters
}
\examples{
abbreviate_middle_names("Franz Jude Wayne")
}
