% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initials.R
\name{add_initials}
\alias{add_initials}
\title{Add initials to the infosheet}
\usage{
add_initials(infosheet)
}
\arguments{
\item{infosheet}{the imported infosheet}
}
\value{
The function returns the infosheet with the initials in
  an additional column
}
\description{
This function adds the initials to the infosheet based on the Firstname,
Middle name and Surname. The function uses whitespaces and hypens to
detect the separate names. Also, the function is case sensitive.
}
