% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_contributors_table.R
\name{check_duplicate_initials}
\alias{check_duplicate_initials}
\title{Check for same initials}
\usage{
check_duplicate_initials(contributors_table)
}
\arguments{
\item{contributors_table}{the imported contributors_table}
}
\value{
The function returns
}
\description{
This function checks the contributors_table for duplicate initials, and
issues a warning that the surnames will be used to differentiate
between the users.
}
