% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_duplicate_initials}
\alias{check_duplicate_initials}
\title{Check for Duplicate Initials}
\usage{
check_duplicate_initials(contributors_table)
}
\arguments{
\item{contributors_table}{A dataframe containing the contributors' information.}
}
\value{
A list containing:
\item{type}{Type of validation result: "success" or "warning".}
\item{message}{An informative message indicating the initials that are duplicated.}
}
\description{
This function checks for duplicate initials in the \code{contributors_table}, taking into
account the \code{Firstname}, \verb{Middle name}, and \code{Surname} columns. It issues a warning
if duplicate initials are found, which may indicate ambiguous contributor identification.
}
