% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_helpers.R
\name{check_missing_order}
\alias{check_missing_order}
\title{Check for missing values in the \verb{Order in publication} column}
\usage{
check_missing_order(contributors_table)
}
\description{
Check for missing values in the \verb{Order in publication} column
}
