% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xml.R
\name{generate_contrib_group}
\alias{generate_contrib_group}
\title{Generate contrib-group XML node}
\usage{
generate_contrib_group(
  contrib_data,
  affiliation_data,
  authors_only,
  contrib_type = "author",
  include_orcid = TRUE
)
}
\arguments{
\item{contrib_data}{preprocessed contributors table with affiliations}

\item{affiliation_data}{vector of unique affiliations}

\item{authors_only}{full authors table for metadata}

\item{contrib_type}{character. Either "author" (default) or "acknowledgee"}

\item{include_orcid}{Logical. If \code{TRUE} (default), includes ORCID IDs in the output.}
}
\description{
Generate contrib-group XML node
}
\keyword{internal}
