% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_yaml.R
\name{print_yaml}
\alias{print_yaml}
\title{Generate an YAML document of the contributions}
\usage{
print_yaml(infosheet)
}
\arguments{
\item{infosheet}{validated infosheet}
}
\value{
The function returns a YAML document
}
\description{
The function generates a YAML document containing the contributors information
and contributions according to the CRediT taxonomy. The output is generated
from an infosheet validated with the \code{\link{validate_infosheet}} function.
The infosheet must be based on the \code{\link{infosheet_template}}.
}
\details{
#' @section Warning:
The function is primarily developed to be the part of a shiny app. As the
  validation is handled inside of the app separately, the function can
  break with non-informative errors if running locally without first
  validating it.
}
\examples{
validate_infosheet(infosheet = infosheet_template)
print_yaml(infosheet = infosheet_template)
}
\seealso{
Other \code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}: 
\code{\link{contrib_affil_print}()},
\code{\link{roles_readable_print}()},
\code{\link{xml_print}()}
}
\concept{\code{\link{print_contrib_affil}}, \code{\link{print_roles_readable}},
  \code{\link{print_xml}}, \code{\link{print_yaml}}}
