% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_yaml.R
\name{print_yaml}
\alias{print_yaml}
\title{Generate an YAML document of the contributions}
\usage{
print_yaml(contributors_table)
}
\arguments{
\item{contributors_table}{validated contributors_table}
}
\value{
The function returns a YAML document
}
\description{
The function generates a YAML document containing the contributors information
and contributions according to the CRediT taxonomy. The output is generated
from an \code{contributors_table} validated with the \code{\link[=validate_contributors_table]{validate_contributors_table()}} function.
The \code{contributors_table} must be based on the \code{\link[=contributors_table_template]{contributors_table_template()}}.
}
\section{Warning}{

The function is primarily developed to be the part of a shiny app. As the
validation is handled inside of the app separately, the function can
break with non-informative errors if running locally without first
validating it.
}

\examples{
example_contributors_table <-
read_contributors_table(
contributors_table = system.file("extdata",
"contributors_table_example.csv", package = "tenzing", mustWork = TRUE))
validate_contributors_table(contributors_table = example_contributors_table)
print_yaml(contributors_table = example_contributors_table)

}
\seealso{
Other output functions: 
\code{\link{print_conflict_statement}()},
\code{\link{print_credit_roles}()},
\code{\link{print_funding}()},
\code{\link{print_title_page}()},
\code{\link{print_xml}()}
}
\concept{output functions}
