% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_contributors_table.R
\name{validate_contributors_table}
\alias{validate_contributors_table}
\title{Validating the contributors table}
\usage{
validate_contributors_table(contributors_table)
}
\arguments{
\item{contributors_table}{dataframe, filled out contributors_table}
}
\value{
The function returns a list for each checked statement. Each list contains
  a \code{type} vector that stores whether the statement passed the check "success"
  or failed "warning" or "error", and a \code{message} vector that contains information
  about the nature of the check.
}
\description{
This function validates the `contributors_table` provided to it by checking whether the
provided `contributors_table` is compatible with the \code{\link{contributors_table_template}}. The function
early escapes only if the provided `contributors_table` is not a dataframe, the variable
names that are present in the `contributors_table_template` are missing, or if the 
`contributors_table` is empty.
}
\section{The function checks the following statements}{

\itemize{
  \item error, the provided contributors_table is a dataframe
  \item error, the provided contributors_table does not have the same column names as the template
  \item error, the provided contributors_table is empty
  \item error, \code{Firstname} variable has missing value for one of the contributors
  \item error, \code{Surname} variable has a missing value for one of the contributors
  \item warning, the contributors_table has duplicate names
  \item warning, the contributors_table has names with duplicate initials
  \item error, the \code{'Order in publication'} variable has missing values
  \item error, the \code{'Order in publication'} variable has duplicate values
  \item error, both \code{'Primary affiliation'} and \code{'Secondary affiliation'} variables
    are missing for one contributor
  \item warning, there is no corresponding author added
  \item warning, email address is missing for the corresponding author
  \item warning, there is at least one CRediT role provided for all contributors
}
}

\examples{
# Read the example contributors table
file_path <- system.file("extdata", "contributors_table_example.csv", package = "tenzing", mustWork = TRUE)
my_contributors_table <- read_contributors_table(contributors_table_path = file_path)
# Validate the table
check_result <- validate_contributors_table(contributors_table = my_contributors_table)
# Show the results of the checks
purrr::map(check_result, "type")
# Show the corresponding messages
purrr::map(check_result, "message")

}
